requirements.txt
pdfplumber==0.10.3
pytesseract==0.3.10
pandas==2.2.2
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
